walgreens.com/requirements.txt
python-dateutil
scrapy
html2text
